Draft : Develop (#5)
merging V2 gui 
---------
