+++ .gitignore
@@ -85,3 +85,6 @@ cython_debug/
 
 ### package ###
 pyproject.toml
+
+### executable packages ###
+ *./output/GUI
