+++ Fiber_Drawing_Tower/Archives/PID.py
@@ -1,0 +1,11 @@
+class PID:
+    def __init__(self, setValue):
+        self.setValue = setValue
+
+    def P(self, P, constanteP):
+        manipulatedVariable = constanteP*(self.setValue-P)
+        return manipulatedVariable
+
+    def PI(self, P, I, constanteP):
+        pass
+
