--- Fiber_Drawing_Tower/CAD_sketches/ToDelete
@@ -1,1 +1,0 @@
-
